refactor: consolidate duplicated utilities into utils.py
Move parse_timestamp, format_duration, and discover_instances from
report.py, campaign.py, and triage.py into lafleur/utils.py. Update
all consumer modules and tests to import from the canonical location.

Closes #798
